requirements.txt
yt-dlp
whisper-openai
youtube-transcript-api
python-dotenv
openai
sqlalchemy
tqdm
ffmpeg-python 